CodeGenerator_ByteCode

+++ AstraNASM/Compiler/Compiler/AST/Nodes/Node_Literal.cs
@@ -19,13 +19,13 @@ public class Node_Literal : Node
         if (constant is Token_Char)
         {
             ctx.gen.Comment($"Literal = '{constant.value}'");
-            result = ctx.gen.Allocate(PrimitiveTypes.LONG);
+            result = ctx.gen.Allocate(PrimitiveTypes.INT);
             ctx.gen.SetValue(result, "'" + constant.value + "'");
         }
         else
         {
             ctx.gen.Comment("Literal = " + constant.value);
-            result = ctx.gen.Allocate(PrimitiveTypes.LONG);
+            result = ctx.gen.Allocate(PrimitiveTypes.INT);
             ctx.gen.SetValue(result, constant.value);
         }
